Move utilities and types to separate directories
