+++ src/core.js
@@ -8,28 +8,25 @@
 import os from 'os';
 import path from 'path';
 import architect from 'architect';
-import { enabledComponents } from './components';
 import * as briskhome from '../package.json';
+import { enabledComponents } from './utilities/components';
+import { briskhomeAsciiLogo } from './utilities/constants';
 
 const writeBriskhomeLogo = (): void => {
   process.stdout.write('\u001b[2J\u001b[0;0H');
-  process.stdout.write(`
-     ██████╗ ██████╗ ██╗███████╗██╗  ██╗██╗  ██╗ ██████╗ ███╗   ███╗███████╗
-     ██╔══██╗██╔══██╗██║██╔════╝██║ ██╔╝██║  ██║██╔═══██╗████╗ ████║██╔════╝
-     ██████╔╝██████╔╝██║███████╗█████╔╝ ███████║██║   ██║██╔████╔██║█████╗
-     ██╔══██╗██╔══██╗██║╚════██║██╔═██╗ ██╔══██║██║   ██║██║╚██╔╝██║██╔══╝
-     ██████╔╝██║  ██║██║███████║██║  ██╗██║  ██║╚██████╔╝██║ ╚═╝ ██║███████╗
-     ╚═════╝ ╚═╝  ╚═╝╚═╝╚══════╝╚═╝  ╚═╝╚═╝  ╚═╝ ╚═════╝ ╚═╝     ╚═╝╚══════╝
-  \n`);
+  process.stdout.write(briskhomeAsciiLogo);
 };
 
 const writeBriskhomeInfo = (): void => {
   process.stdout.write(`{"name":"briskhome","hostname":"${os.hostname()}","pid":${process.pid},"component":"core","level":30,"msg":"Initializing Briskhome v${briskhome.version}","time":"${new Date().toISOString()}","v":0}\n`);
+  process.title = 'briskhome';
 };
 
-process.title = 'briskhome';
-writeBriskhomeLogo();
-writeBriskhomeInfo();
+// TODO: Need a better name
+if (!process.argv.includes('--ugly')) {
+  writeBriskhomeLogo();
+  writeBriskhomeInfo();
+}
 
 const components = architect.resolveConfig(enabledComponents(), path.resolve(__dirname));
