+++ src/core.log/index.js
@@ -5,7 +5,7 @@
 
 import bunyan from 'bunyan';
 import { getCallee } from '../utilities/helpers';
-import type { CoreImports, CoreRegister } from '../utilities/coreTypes';
+import type { CoreImports, CoreRegister } from '../types/coreTypes';
 
 export default (options: Object, imports: CoreImports, register: CoreRegister) => {
   const config = imports.config();
